refactor: made changes to the folder structure; now using src structure
