+++ src/app/onboarding/components/IndustrySelection.tsx
