+++ src/app/onboarding/components/JobSearchStatus.tsx
