+++ src/app/onboarding/components/SkillsSelection.tsx
